// components/event-calendar/types.ts
export type CalendarView = "month" | "week" | "day" | "agenda" | "year" | "schedule" | "4day";

export interface CalendarEvent {
  id: string;
  title: string;
  description?: string;
  start: Date;
  end: Date;
  allDay?: boolean;
  color?: EventColor;
  label?: string;
  location?: string;
}

export type EventColor = "blue" | "orange" | "violet" | "rose" | "emerald" | "sky" | "indigo" | "purple" | "pink" | "red" | "yellow" | "green" | "teal" | "cyan";